% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCoords.R
\name{convertCoords}
\alias{convertCoords}
\title{convertCoords}
\usage{
convertCoords(inputData, transformMatrix, colNames = c("x", "y"))
}
\arguments{
\item{inputData}{dataframe with coordinates in the arbitrary coordinate system}

\item{transformMatrix}{transformation matrix, produced by `getTransform`.}

\item{colNames}{vector of length 2. Name of columns that will receive data.}
}
\value{
dataframe with coordinates converted to the destination coordinate system.
}
\description{
Converts coordinates between coordinate systems, using a transform matrix produced by `getTransform`
}
\examples{
convertCoords
}
