% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datathief-package.R
\docType{package}
\name{datathief-package}
\alias{datathief-package}
\title{datathief: Get Data From Screenshots of Figures}
\description{
Package provides data scraping tools for images where the underlying data is not readily accessible, such as a screenshot of a powerpoint slide, an old manuscript, gray literature, and so on.
}
\author{
\strong{Maintainer}: Troy Hill \email{Hill.Troy@gmail.com}

}
\keyword{internal}
