% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPoints.R
\name{getPoints}
\alias{getPoints}
\title{getPoints}
\usage{
getPoints(locations = 3, type = "known")
}
\arguments{
\item{locations}{number of points to be selected}

\item{type}{the type of points you will be selecting. Can be either `known` or `unknown`. If `known`, the user is asked for coordinates after selecting each point.}
}
\value{
a dataframe of xy coordinates. If `type = known`, the dataframe can be used to calculate a matrix transformation to reconcile two coordinate systems.
}
\description{
Function lets the user identify points on a displayed plot.
}
\details{
TBD
}
\examples{
getPoints

}
