% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTransform.R
\name{getTransform}
\alias{getTransform}
\title{getTransform}
\usage{
getTransform(
  knownData,
  x.dest = "x_true",
  y.dest = "y_true",
  x.arb = "x_arb",
  y.arb = "y_arb",
  n = 3
)
}
\arguments{
\item{knownData}{a dataframe with data points with known locations in both coordinate systems, created by getPoints}

\item{x.dest}{name of column with x coordinates in destination coordinate system}

\item{y.dest}{name of column with y coordinates in destination coordinate system}

\item{x.arb}{name of column with x coordinates in the arbitrary coordinate system}

\item{y.arb}{name of column with y coordinates in the arbitrary coordinate system}

\item{n}{a number. don't adjust this.}
}
\value{
a matrix
}
\description{
Generates a transform function that converts between two coordinate systems.
}
\examples{
getTransform
}
