% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDialogBox.R
\name{makeDialogBox}
\alias{makeDialogBox}
\title{makeDialogBox}
\usage{
makeDialogBox(
  vars =
    c("How many additional unkown data points are needed? Enter `0` if you are done."),
  title = "Variable Entry",
  prompt = NULL
)
}
\arguments{
\item{vars}{text appearing in popup box}

\item{title}{NULL}

\item{prompt}{NULL}
}
\value{
value input by the user into the dialog box.
}
\description{
Creates a dialog box to ask if the user wants to input more points.
}
\examples{
makeDialogBox
}
